add unit test for product service
