+++ src/main/kotlin/com/example/ordrsmartapi/utils/mapper/IEntityDtoMapper.kt
@@ -5,11 +5,8 @@ import com.example.ordrsmartapi.dto.AddressDTO
 import com.example.ordrsmartapi.dto.ProductDTO
 import com.example.ordrsmartapi.entity.Address
 import com.example.ordrsmartapi.entity.Product
-import org.springframework.stereotype.Repository
-import org.springframework.stereotype.Service
 
 
-@Repository
 interface IEntityDtoMapper {
     fun fromProductEntityToDto(entity: Product): ProductDTO
     fun fromProductDtoToEntity(domain: ProductDTO): Product
