--- src/test/kotlin/com/example/ordrsmartapi/SampleTest.kt
@@ -1,14 +1,0 @@
-package com.example.ordrsmartapi
-
-import org.junit.jupiter.api.Test
-
-import org.junit.jupiter.api.Assertions.*
-
-class SampleTest {
-    private val testSample: Sample = Sample()
-    @Test
-    fun sum() {
-        val expected = 42
-        assertEquals(expected, testSample.sum(40, 2))
-    }
-}
